+++ src/base/string/random.ts
@@ -5,3 +5,5 @@ export function randomString(e = 32) {
   for (let i = 0; i < e; i++) n += t.charAt(Math.floor(Math.random() * a));
   return n;
 }
+
+export const getRandomString = randomString;