+++ src/gray-release/types.ts
@@ -1,0 +1,1 @@
+export type IGlobalGrayPublishConfig = Record<string, Record<string, string>>;