+++ src/mp/index.ts
@@ -1,2 +1,3 @@
+export { getComponentInfo, getPageTotal } from './page-total';
 export { saveMpImage } from './save-mp-image';
 export { updateManager } from './update-manager';