+++ src/time/index.ts
@@ -1,1 +1,2 @@
+export { getNowBySecond } from './now';
 export { timeStampFormat, dateFormat } from './time';