+++ src/time/now.ts
@@ -1,0 +1,3 @@
+export function getNowBySecond() {
+  return Math.floor(Date.now() / 1000);
+}