+++ docs/CHANGELOG.md
@@ -1,5 +1,62 @@
 # 更新日志
 
+### [1.3.150](https://github.com/novlan1/t-comm/compare/v1.3.149...v1.3.150) (2025-03-08)
+
+
+### Features 🎉
+
+* **git:** add checkGitClean ([fc0aa1d](https://github.com/novlan1/t-comm/commit/fc0aa1d64b5499dc39e333092ccc3b88d19173ea))
+
+### [1.3.149](https://github.com/novlan1/t-comm/compare/v1.3.148...v1.3.149) (2025-02-24)
+
+
+### Features 🎉
+
+* **minimatch:** add getAuditorFromRainbowConfig ([29b5f34](https://github.com/novlan1/t-comm/commit/29b5f34f887d227175a85bc5f256de2ec8c78ba1))
+
+### [1.3.148](https://github.com/novlan1/t-comm/compare/v1.3.147...v1.3.148) (2025-02-21)
+
+
+### Features 🎉
+
+* **gray-release:** add gray-release ([24a04a5](https://github.com/novlan1/t-comm/commit/24a04a5f1f2834905be45ff8015e88a42ce7120a))
+
+### [1.3.147](https://github.com/novlan1/t-comm/compare/v1.3.146...v1.3.147) (2025-02-13)
+
+
+### Features 🎉
+
+* **pipeline:** 支持mentionList ([2e7761f](https://github.com/novlan1/t-comm/commit/2e7761f8dcaed0c978b301a1b3bb690330b92c34))
+
+### [1.3.146](https://github.com/novlan1/t-comm/compare/v1.3.145...v1.3.146) (2025-02-11)
+
+
+### Features 🎉
+
+* **mp:** 优化getComponentInfo ([ab1ed66](https://github.com/novlan1/t-comm/commit/ab1ed66743389ab45a132510f4d6089757d1dbfc))
+
+### [1.3.145](https://github.com/novlan1/t-comm/compare/v1.3.144...v1.3.145) (2025-02-11)
+
+
+### Features 🎉
+
+* **mp:** add getPageTotal ([66542ed](https://github.com/novlan1/t-comm/commit/66542ed2c8bb640304a15b2f6698680ee0479696))
+* **rum:** 增强健壮性 ([388bb71](https://github.com/novlan1/t-comm/commit/388bb711d5231915f4297e3d7a69fd795de88b35))
+* **time:** add now time ([56cab45](https://github.com/novlan1/t-comm/commit/56cab453edbe60c3892b698180bd27ff8667c1cf))
+
+### [1.3.144](https://github.com/novlan1/t-comm/compare/v1.3.143...v1.3.144) (2025-01-18)
+
+
+### 👷 Continuous Integration | CI 配置
+
+* update workflow ([4bf882d](https://github.com/novlan1/t-comm/commit/4bf882d5154d8e534329268296b9dc78810be85b))
+
+
+### Features 🎉
+
+* **lint:** 优化lint ([a4769f1](https://github.com/novlan1/t-comm/commit/a4769f1066b2678c0f2878307c308192c6ef1345))
+* update lint ([2392c91](https://github.com/novlan1/t-comm/commit/2392c9133899838f01f148975593a6ae0a3c6f16))
+
 ### [1.3.143](https://github.com/novlan1/t-comm/compare/v1.3.142...v1.3.143) (2025-01-07)